# cmd: fix completion of remotes